+++ src/app/[variants]/(main)/chat/(workspace)/@conversation/features/ChatList/WelcomeChatItem/AgentWelcome/OpeningQuestions.tsx
@@ -8,7 +8,7 @@ import { Flexbox } from 'react-layout-kit';
 
 import { useChatStore } from '@/store/chat';
 
-import { useSend } from '../../ChatInput/useSend';
+import { useSend } from '../../../ChatInput/useSend';
 
 const useStyles = createStyles(({ css, token, responsive }) => ({
   card: css`
