--- src/app/[variants]/(main)/chat/(workspace)/@conversation/features/ChatList/WelcomeChatItem/WelcomeMessage.tsx
@@ -1,65 +1,0 @@
-import isEqual from 'fast-deep-equal';
-import qs from 'query-string';
-import React, { useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Flexbox } from 'react-layout-kit';
-
-import { ChatItem } from '@/features/ChatItem';
-import { useAgentStore } from '@/store/agent';
-import { agentChatConfigSelectors, agentSelectors } from '@/store/agent/selectors';
-import { useChatStore } from '@/store/chat';
-import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
-import { useSessionStore } from '@/store/session';
-import { sessionMetaSelectors } from '@/store/session/selectors';
-
-import OpeningQuestions from './OpeningQuestions';
-
-const WelcomeMessage = () => {
-  const mobile = useServerConfigStore((s) => s.isMobile);
-  const { t } = useTranslation('chat');
-  const type = useAgentStore(agentChatConfigSelectors.displayMode);
-  const openingMessage = useAgentStore(agentSelectors.openingMessage);
-  const openingQuestions = useAgentStore(agentSelectors.openingQuestions);
-
-  const meta = useSessionStore(sessionMetaSelectors.currentAgentMeta, isEqual);
-  const { isAgentEditable } = useServerConfigStore(featureFlagsSelectors);
-  const activeId = useChatStore((s) => s.activeId);
-
-  const agentSystemRoleMsg = t('agentDefaultMessageWithSystemRole', {
-    name: meta.title || t('defaultAgent'),
-    systemRole: meta.description,
-  });
-
-  const agentMsg = t(isAgentEditable ? 'agentDefaultMessage' : 'agentDefaultMessageWithoutEdit', {
-    name: meta.title || t('defaultAgent'),
-    url: qs.stringifyUrl({
-      query: mobile ? { session: activeId, showMobileWorkspace: mobile } : { session: activeId },
-      url: '/chat/settings',
-    }),
-  });
-
-  const message = useMemo(() => {
-    if (openingMessage) return openingMessage;
-    return !!meta.description ? agentSystemRoleMsg : agentMsg;
-  }, [openingMessage, agentSystemRoleMsg, agentMsg, meta.description]);
-
-  const chatItem = (
-    <ChatItem
-      avatar={meta}
-      editing={false}
-      message={message}
-      placement={'left'}
-      variant={type === 'chat' ? 'bubble' : 'docs'}
-    />
-  );
-
-  return openingQuestions.length > 0 ? (
-    <Flexbox>
-      {chatItem}
-      <OpeningQuestions mobile={mobile} questions={openingQuestions} />
-    </Flexbox>
-  ) : (
-    chatItem
-  );
-};
-export default WelcomeMessage;
